test: add initial controller tests for category

+++ app/controllers/category_controller.py
@@ -27,23 +27,26 @@ class CategoryController:
             raise ControllerException(detail=e) from e
         return results
 
-    async def add_category(self, category) -> Category:
+    async def add_category(self, category: Category) -> Category:
         """
         Creates a new category in the database
 
         Args:
             category_name (str): The name of the new category
         """
-
+        if not isinstance(category, Category):
+            raise ValueError("Not of correct type")
         try:
-            category = await self.category_service.create(category)
+            added_category = await self.category_service.create(category)
         except ValueError as e:
+            raise ValueError(e) from e
+        except ServiceException as e:
             raise ControllerException(detail=e) from e
-        except Exception as e:
-            raise ControllerException(detail=e) from e
-        return category
+        return added_category
 
     async def remove_category(self, category_id: UUID):
+        if not isinstance(category_id, UUID):
+            raise ValueError("Invalid Category Id")
         try:
             await self.category_service.delete(category_id)
         except ServiceException as e:
@@ -60,11 +63,13 @@ class CategoryController:
         Args:
             category (Category): _description_
         """
+
         try:
             category = await self.category_service.put(category.id, category)
         except ServiceException as e:
+            message = e
             raise ControllerException(
-                detail=f"There was an issue with the category service when updating a category:\n {e}"
+                detail=f"There is an issue with the category service: {message}"
             ) from e
         except Exception as e:
             raise ControllerException(detail=e) from e
